requirements.txt
torch>=2.0.0
torchvision
cupy-cuda12x
transformers==4.56.0
controlnet-aux==0.0.7
imageio
imageio[ffmpeg]
safetensors
einops
sentencepiece
protobuf
modelscope
ftfy
pynvml
pandas
accelerate
qwen_vl_utils
peft==0.17.0
lightning
fairscale
loralib
pillow_heif
matplotlib
gradio
lpips